Integrating Min Cost Flow
making min cost flow work more elegantly with our graph architecture.

+++ src/oarlib/core/Link.java
@@ -14,9 +14,9 @@ public abstract class Link<V extends Vertex> {
 	private int guid; //the idea is that this will be unique for all links, even between graphs
 	private int mId; //while this will help us identify the 'same' link in different graphs (graph copies for instance)
 	private Pair<V> mEndpoints;
-	private double mCost;
+	private int mCost;
 	
-	public Link(String label, Pair<V> endpoints, double cost)
+	public Link(String label, Pair<V> endpoints, int cost)
 	{
 		//if mId gets set to a negative number, then we know it hasn't been set.
 		setmId(-1);
@@ -26,7 +26,7 @@ public abstract class Link<V extends Vertex> {
 		setCost(cost);
 		counter++;
 	}
-	public Link(String label, Pair<V> endpoints, double cost, int id) throws IllegalArgumentException
+	public Link(String label, Pair<V> endpoints, int cost, int id) throws IllegalArgumentException
 	{
 		if( id < 0)
 			throw new IllegalArgumentException("negative ids are reserved for detecting when ids have not been set.");
@@ -66,11 +66,11 @@ public abstract class Link<V extends Vertex> {
 		this.mEndpoints = mEndpoints;
 	}
 
-	public double getCost() {
+	public int getCost() {
 		return mCost;
 	}
 
-	public void setCost(double mCost) {
+	public void setCost(int mCost) {
 		this.mCost = mCost;
 	}
 	public int getmId() {
